+++ go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/DATA-DOG/go-sqlmock v1.3.3
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/gofrs/flock v0.7.1
-	github.com/gogo/protobuf v1.3.0 // indirect
+	github.com/gogo/protobuf v1.3.0
 	github.com/rs/xid v1.2.1
 	go.uber.org/zap v1.10.0
 	google.golang.org/grpc v1.23.0
